backend/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# LangChain & LangGraph
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.5
langchain-community==0.0.10

# Database
sqlalchemy==2.0.23
alembic==1.13.0
aiosqlite==0.19.0

# Vector Database
qdrant-client==1.7.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
pydantic-settings==2.1.0

# API & Networking
httpx==0.25.2
websockets==12.0
sse-starlette==1.8.2

# Utilities
numpy==1.24.3
tiktoken==0.5.2
tenacity==8.2.3
python-dateutil==2.8.2

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0